2.x: Subject/Processor improvements & small cleanup (#4437)

+++ src/test/java/io/reactivex/internal/operators/single/SingleDoOnTest.java
@@ -79,7 +79,7 @@ public class SingleDoOnTest {
         })
         .test();
         
-        assertEquals((Integer)1, event[0]);
+        assertEquals(1, event[0]);
     }
 
 }
